Split parsing into modules and material list fixes
