+++ litematica_tools/__init__.py
@@ -1,2 +1,3 @@
 from .material_list import MaterialList, Counter
-from .schematic_parse import NBT_File, Region
+from .schematic_parse import NBTFile
+from .shared_storage import Structure, Region
